조회수 증가 단위 테스트

+++ back-end/src/main/java/com/webservice/bookstore/service/ItemService.java
@@ -32,7 +32,7 @@ public class ItemService {
     }
 
     @Transactional
-    public void improveViewCount(int id) {
+    public void improveViewCount(Long id) {
         this.itemRepository.improveViewCount(id);
     }
 }
